+++ www/src/views/system/menus/ButtonForm.vue
@@ -2,8 +2,11 @@
   <el-form-item label="按钮:" prop="name" :rules="[{required:true,message:'请填写按钮名称'}]">
     <el-input v-model="selectData.name"></el-input>
   </el-form-item>
-  <el-form-item label="上级菜单" :rules="[{required:true,message:'请选择上级菜单'}]">
-    <el-input v-model="selectData.parent_id"></el-input>
+  <el-form-item label="上级菜单" prop="parent_id" :rules="[{required:true,message:'请选择上级菜单'}]">
+    <el-cascader v-model="selectData.parent_id" :options="cascaderMenu"
+                 :props="{checkStrictly:true,value:'id',label:'name',emitPath:false}"
+                 placeholder="请选择父菜单" style="width:100%">
+    </el-cascader>
   </el-form-item>
   <el-form-item label="授权标识" prop="path" :rules="[{required:true,message:'请填写按钮授权标识'}]">
     <el-input v-model="selectData.path"></el-input>
@@ -24,7 +27,8 @@
 </template>
 
 <script>
-  import {toRefs, reactive, ref, watch, inject} from 'vue'
+  import {toRefs, reactive, ref, inject} from 'vue'
+  import useMenu from '@/composables/useMenu'
 
   export default {
     name: 'ButtonForm',
@@ -32,32 +36,15 @@
     emits: ['update:form'],
     setup(props, {emit}) {
       const {form} = toRefs(props)
-      const selectData = reactive(form.value)
       const loading = ref(false)
-      const selectApis = ref([])
 
       const menuData = inject('menuData')
-      if(selectData.id !== null){
-        for (const api of selectData.apis) {
-          console.log(api)
-          selectApis.value.push(api.id)
-        }
-      }
-
-
-      watch(selectData, (newData) => {
-        console.log('watch selectData change:' + selectData)
-        emit('update:form', selectData)
-      })
-
-      watch(selectApis,(newValue)=>{
-        console.log('selectApi watch')
-        selectData.apis = selectApis.value
-      })
+      const {selectData, selectApis, cascaderMenu} = useMenu(form.value, menuData.value, emit)
 
       return {
         selectData,
         menuData,
+        cascaderMenu,
         loading,
         selectApis
       }
